table: split auto-merge tests into new file
